+++ synapse-core/src/main/java/de/otto/synapse/annotation/BeanNameHelper.java
@@ -13,12 +13,12 @@ public class BeanNameHelper {
         return LOWER_HYPHEN.to(LOWER_CAMEL, channelName) + "MessageLogReceiverEndpoint";
     }
 
-    public static String beanNameForMessageQueue(final String channelName) {
-        return LOWER_HYPHEN.to(LOWER_CAMEL, channelName) + "MessageQueue";
-    }
-
     public static String beanNameForMessageQueueReceiverEndpoint(final String channelName) {
         return LOWER_HYPHEN.to(LOWER_CAMEL, channelName) + "MessageQueueReceiverEndpoint";
+    }
+
+    public static String beanNameForMessageQueueSenderEndpoint(final String channelName) {
+        return LOWER_HYPHEN.to(LOWER_CAMEL, channelName) + "MessageQueueSenderEndpoint";
     }
 
 }