+++ synapse-core/src/main/java/de/otto/synapse/annotation/messagequeue/EnableMessageQueueSenderEndpoints.java
@@ -1,0 +1,16 @@
+package de.otto.synapse.annotation.messagequeue;
+
+import de.otto.synapse.configuration.SynapseAutoConfiguration;
+import org.springframework.context.annotation.Import;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Import({
+        SynapseAutoConfiguration.class,
+        MessageQueueSenderEndpointBeanRegistrar.class})
+public @interface EnableMessageQueueSenderEndpoints {
+    EnableMessageQueueSenderEndpoint[] value();
+}